Plan for NVDA Data Pipeline
