+++ .gitignore
@@ -23,3 +23,6 @@ wheels/
 .user.yml
 profiles.yml
 logs
+
+# Data source
+data/
