+++ src/data_processing/modules/__init__.py
@@ -1,2 +1,3 @@
-from src.data_processing.modules.load_pdf import asset1, asset2
+from src.data_processing.modules.documents_llms.load_pdf import pdf_embedding, pdf_loader, pdf_extraction, pdf_path, pdf_vector_database
+from src.data_processing.modules.extract.NVDA import ingest_data
